Refactors components, adds styling to the error messages
