+++ client/bundles/Main/components/Main.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import MainSessionsDisplay from "./MainSessionsDisplay";
+import MainSessionsDisplay from "./sessions/MainSessionsDisplay";
 import MainNavLinks from "./MainNavLinks";
 
 
@@ -46,23 +46,12 @@ export default class Main extends React.Component {
 					<img src="./assets/logo.png" id="logo" alt="logo" />
 					<p id="title">Tribal</p>
 				</div>
-				<div className='wrapper-col content'>
-				</div>
-				<div className='content'>
-					<div>
-						<div className="wrapper-col">
-							<MainNavLinks user={this.state.user} logout={this.logout}/>
-							<h4 className="m-4">Hello, {this.state.user ? this.state.user.first_name : 'Stranger'}!</h4>
-						</div>
-					</div>
+				<div className='content wrapper-col'>
+					<MainNavLinks user={this.state.user} logout={this.logout}/>
+					<h4 className="m-4">Hello, {this.state.user ? this.state.user.first_name : 'Stranger'}!</h4>
 					<div style = {this.state.user? {}:{display: "none"} }>
 						<MainSessionsDisplay  sessions={this.state.sessions} />
 					</div>
-					
-				</div>
-
-				<div className='footer'>
-					<p>FOOTER</p>
 				</div>
 			</div>
 		);
