+++ client/bundles/Main/components/MainNavLinks.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { NavLink, Link } from "react-router-dom";
+import { NavLink } from "react-router-dom";
 
 export class MainNavLinks extends Component {
 	render() {
