--- client/bundles/Main/components/WithRouter.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { withRouter } from "react-router-dom";
-
-
-
-export default function withRouter() {
-    this.props.history.push("/some/Path");
-}
