+++ client/bundles/Main/components/authentication/Registration.jsx
@@ -21,26 +21,21 @@ export class Registration extends Component {
 		this.onSubmit = this.onSubmit.bind(this);
 		this.onRoleClick = this.onRoleClick.bind(this);
 	}
-
 	onSubmit() {
 		event.preventDefault();
 		const csrfToken = ReactOnRails.authenticityToken();
 		let registrationForm = document.getElementById("Registration-form");
 		const data = new FormData(registrationForm);
-
 		const config = {
 			headers: {
 				"Content-Type": "application/x-www-form-urlencoded",
 				"X-CSRF-Token": csrfToken
 			}
 		};
-
 		axios
 			.post("/users", data, config)
 			.then(response => {
-
 				if (response.data.errors) {
-					debugger;
 					let errors = Object.entries(response.data.errors).join('\n').replace(/,|_/g, ' ');
 					this.setState({
 						formErrors: errors
@@ -48,14 +43,11 @@ export class Registration extends Component {
 				} else {
 					document.location.href = "/";
 				}
-
-
 			})
 			.catch(function (error) {
 				console.log(error);
 			});
 	}
-
 	onClick(event) {
 		event.preventDefault();
 		this.setState({
@@ -92,7 +84,7 @@ export class Registration extends Component {
 					</div>
 				</div>
 				<div style={this.state.roleChoiceStatus ? {} : { display: "none" }}>
-					<div style={{ whiteSpace: 'pre-wrap' }}>{this.state.formErrors}</div>
+					<div className="mt-4 text-center" style={{ whiteSpace: 'pre-wrap' }}>{this.state.formErrors}</div>
 					<RegistrationForm
 						onSubmit={this.onSubmit}
 						onChange={this.onChange}
