+++ client/bundles/Main/components/authentication/RegistrationForm.jsx
