+++ client/bundles/Main/components/sessions/AllSessions.jsx
